Refactor classes into appropriate packages
