+++ src/main/java/duke/dukeExceptions/DukeException.java
@@ -1,3 +1,5 @@
+package duke.dukeExceptions;
+
 public class DukeException extends Exception{
     public DukeException(String errorMsg) {
         super("OPS!!!! " + errorMsg);
