+++ src/main/java/duke/tasks/Deadline.java
@@ -1,15 +1,18 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import duke.dukeExceptions.DukeException;
 
-public class Event extends Task {
+public class Deadline extends Task {
     private LocalDate d;
     private LocalTime t;
     private boolean hasTime;
 
-    Event(String description, String at) throws DukeException {
+    public Deadline(String description, String by) throws DukeException {
         super(description);
-        String[] dateAndTime = at.split(" ");
+        String[] dateAndTime = by.split(" ");
         try {
             this.d = LocalDate.parse(dateAndTime[0]);
             if (dateAndTime.length == 2) {
@@ -19,31 +22,31 @@ public class Event extends Task {
                 this.hasTime = false;
             }
         } catch (java.time.format.DateTimeParseException e) {
-            throw new DukeException("Please enter a valid date and time YYYY-MM-DD HHMM");
+            throw new DukeException("Please enter a valid date and time (YYYY-MM-DD HHMM) or date only (YYYY-MM-DD)");
         }
     }
 
-    public static Event taskFromSave(String saveString) throws DukeException {
+    public static Deadline taskFromSave(String saveString) throws DukeException {
         String[] tokens = saveString.split(" \\| ");
-        String time = tokens[3].equals("true") ? tokens[4] + " " + tokens[5] : tokens[4];
-        Event event = new Event(tokens[2], time);
+        String time = tokens[3].equals("true") ? tokens[4] + " " + tokens[5] : tokens[4]; 
+        Deadline deadline = new Deadline(tokens[2], time);
         if (tokens[1].equals("1")) {
-            event.markDone();
+            deadline.markDone();
         }
-        return event;
-    }
-
-    @Override
-    public String saveString() {
-        return "E | " + super.saveString() +  " | " + this.hasTime + " | " + this.d + " | " 
-                + (this.hasTime ? this.t.format(DateTimeFormatter.ofPattern("HHmm")) : "");
+        return deadline;
     }
     
     @Override
+    public String saveString() {
+        return "D | " + super.saveString() + " | " + this.hasTime + " | " + this.d + " | " 
+                + (this.hasTime ? this.t.format(DateTimeFormatter.ofPattern("HHmm")) : "");
+    }
+
+    @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " 
-                + this.d.format(DateTimeFormatter.ofPattern(("MMM d yyyy")))
-                + (this.hasTime ? " " + this.t : "")  
+        return "[D]" + super.toString() + " (by: " 
+                + this.d.format(DateTimeFormatter.ofPattern(("MMM d yyyy"))) 
+                + (this.hasTime ? " " + this.t : "") 
                 + ")";
     }
 }
