+++ src/main/java/duke/tasks/Event.java
@@ -1,18 +1,19 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
+import duke.dukeExceptions.DukeException;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private LocalDate d;
     private LocalTime t;
     private boolean hasTime;
 
-    Deadline(String description, String by) throws DukeException {
+    public Event(String description, String at) throws DukeException {
         super(description);
-        String[] dateAndTime = by.split(" ");
+        String[] dateAndTime = at.split(" ");
         try {
-            System.out.println(dateAndTime[0]);
             this.d = LocalDate.parse(dateAndTime[0]);
             if (dateAndTime.length == 2) {
                 this.t = LocalTime.parse(dateAndTime[1], DateTimeFormatter.ofPattern("HHmm"));
@@ -21,31 +22,31 @@ public class Deadline extends Task {
                 this.hasTime = false;
             }
         } catch (java.time.format.DateTimeParseException e) {
-            throw new DukeException("Please enter a valid date and time (YYYY-MM-DD HHMM) or date only (YYYY-MM-DD)");
+            throw new DukeException("Please enter a valid date and time YYYY-MM-DD HHMM");
         }
     }
 
-    public static Deadline taskFromSave(String saveString) throws DukeException {
+    public static Event taskFromSave(String saveString) throws DukeException {
         String[] tokens = saveString.split(" \\| ");
-        String time = tokens[3].equals("true") ? tokens[4] + " " + tokens[5] : tokens[4]; 
-        Deadline deadline = new Deadline(tokens[2], time);
+        String time = tokens[3].equals("true") ? tokens[4] + " " + tokens[5] : tokens[4];
+        Event event = new Event(tokens[2], time);
         if (tokens[1].equals("1")) {
-            deadline.markDone();
+            event.markDone();
         }
-        return deadline;
+        return event;
+    }
+
+    @Override
+    public String saveString() {
+        return "E | " + super.saveString() +  " | " + this.hasTime + " | " + this.d + " | " 
+                + (this.hasTime ? this.t.format(DateTimeFormatter.ofPattern("HHmm")) : "");
     }
     
     @Override
-    public String saveString() {
-        return "D | " + super.saveString() + " | " + this.hasTime + " | " + this.d + " | " 
-                + (this.hasTime ? this.t.format(DateTimeFormatter.ofPattern("HHmm")) : "");
-    }
-
-    @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " 
-                + this.d.format(DateTimeFormatter.ofPattern(("MMM d yyyy"))) 
-                + (this.hasTime ? " " + this.t : "") 
+        return "[E]" + super.toString() + " (at: " 
+                + this.d.format(DateTimeFormatter.ofPattern(("MMM d yyyy")))
+                + (this.hasTime ? " " + this.t : "")  
                 + ")";
     }
 }
