+++ src/main/java/duke/tasks/ToDo.java
@@ -1,5 +1,7 @@
+package duke.tasks;
+
 public class ToDo extends Task {
-    ToDo(String description) {
+    public ToDo(String description) {
         super(description);
     }
