+++ text-ui-test/EXPECTED.TXT
@@ -11,47 +11,47 @@ ____________________________________________________________
 	
 ____________________________________________________________
 ____________________________________________________________
-	I've added: [T][ ] Duke you have 1 tasks left
+	I've added: [T][ ] duke you have 1 tasks left
 ____________________________________________________________
  2022-01-02 2359
 2022-01-02
 ____________________________________________________________
-	I've added: [D][ ] do Duke (by: Jan 2 2022 23:59) you have 2 tasks left
+	I've added: [D][ ] do duke (by: Jan 2 2022 23:59) you have 2 tasks left
 ____________________________________________________________
 ____________________________________________________________
-	1 [T][ ] Duke
-	2 [D][ ] do Duke (by: Jan 2 2022 23:59)
+	1 [T][ ] duke
+	2 [D][ ] do duke (by: Jan 2 2022 23:59)
 ____________________________________________________________
 ____________________________________________________________
 	I've added: [E][ ] CS2103T lecture (at: Aug 27 2022) you have 3 tasks left
 ____________________________________________________________
 ____________________________________________________________
-	1 [T][ ] Duke
-	2 [D][ ] do Duke (by: Jan 2 2022 23:59)
+	1 [T][ ] duke
+	2 [D][ ] do duke (by: Jan 2 2022 23:59)
 	3 [E][ ] CS2103T lecture (at: Aug 27 2022)
 ____________________________________________________________
 ____________________________________________________________
-	I have marked: [T][X] Duke as done
+	I have marked: [T][X] duke as done
 ____________________________________________________________
 ____________________________________________________________
-	1 [T][X] Duke
-	2 [D][ ] do Duke (by: Jan 2 2022 23:59)
+	1 [T][X] duke
+	2 [D][ ] do duke (by: Jan 2 2022 23:59)
 	3 [E][ ] CS2103T lecture (at: Aug 27 2022)
 ____________________________________________________________
 ____________________________________________________________
-	I have marked: [D][X] do Duke (by: Jan 2 2022 23:59) as done
+	I have marked: [D][X] do duke (by: Jan 2 2022 23:59) as done
 ____________________________________________________________
 ____________________________________________________________
-	1 [T][X] Duke
-	2 [D][X] do Duke (by: Jan 2 2022 23:59)
+	1 [T][X] duke
+	2 [D][X] do duke (by: Jan 2 2022 23:59)
 	3 [E][ ] CS2103T lecture (at: Aug 27 2022)
 ____________________________________________________________
 ____________________________________________________________
-	I have marked: [D][ ] do Duke (by: Jan 2 2022 23:59) as not done
+	I have marked: [D][ ] do duke (by: Jan 2 2022 23:59) as not done
 ____________________________________________________________
 ____________________________________________________________
-	1 [T][X] Duke
-	2 [D][ ] do Duke (by: Jan 2 2022 23:59)
+	1 [T][X] duke
+	2 [D][ ] do duke (by: Jan 2 2022 23:59)
 	3 [E][ ] CS2103T lecture (at: Aug 27 2022)
 ____________________________________________________________
 ____________________________________________________________
