+++ text-ui-test/input.txt
@@ -1,6 +1,6 @@
 list
-todo Duke
-deadline do Duke /by 2022-01-02 2359
+todo duke
+deadline do duke /by 2022-01-02 2359
 list
 event CS2103T lecture /at 2022-08-27
 list
